mcp-server/pyproject.toml
[project]
name = "mcp-code-execution-server"
version = "0.1.0"
description = "Code execution server using Model Context Protocol"
requires-python = ">=3.10"
dependencies = [
    "mcp-python-sdk>=0.1.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.2",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"] 